feat(crank): Add Start+Select option for dock/undock events
Extends the crank's dock and undock event handlers to support a new 'Start+Select' action.

This change updates the logic in `game_scene.c` to check for `PREF_BUTTON_START_SELECT` and set the `button_hold_mode` to 3, allowing users to trigger both buttons simultaneously when the crank is docked or undocked.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>