feat(script/links_awakening): Map save menu to crank dock action
This change improves the user experience for the Link's Awakening script by making the save menu more intuitive to access.

Previously, opening the save menu required a "Start+Select" input. This has been remapped to the simple, physical action of docking the crank.

The script now forces the `crank_dock_button` preference to `PREF_BUTTON_START_SELECT` and updates the feature description to clearly communicate this new, more streamlined interaction to the user.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>