feat(download): Add prompt to enable patches after download
This commit improves the user workflow by replacing the static "Patch downloaded" message with an interactive modal.

After a patch is successfully downloaded, the user is now asked if they would like to enable it immediately. This streamlines the process, removing the need for the user to manually navigate to the "Manage Patches" screen to activate it.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>